# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Display Router API
  version: 0.1.0
  description: |
    Authentication:
    - Client API uses X-API-Key per client.
    - Admin API uses Authorization: Bearer <ADMIN_TOKEN>.
    - Display server uses X-Display-Secret for /display/health and /display/ws.
servers:
  - url: https://router.example.com
security:
  - ClientApiKey: []
  - AdminBearerAuth: []
paths:
  /api/clients:
    post:
      summary: Register client
      tags: [client-admin]
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '200':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/clients/{client_id}:
    get:
      summary: Get client
      tags: [client-admin]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/clients/{client_id}/payload-types:
    get:
      summary: List client payload types
      tags: [client]
      security:
        - ClientApiKey: []
        - AdminBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Payload types
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayloadTypeList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/payloads:
    post:
      summary: Submit payload
      tags: [client]
      security:
        - ClientApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayloadSubmit'
      responses:
        '200':
          description: Payload accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayloadAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/preview:
    post:
      summary: Preview payload on admin sim display
      tags: [client]
      security:
        - ClientApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewSubmit'
      responses:
        '200':
          description: Preview accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/validate:
    post:
      summary: Validate payload or template
      tags: [client]
      security:
        - ClientApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/templates:
    get:
      summary: List templates
      tags: [client]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      security:
        - ClientApiKey: []
        - AdminBearerAuth: []
      responses:
        '200':
          description: Templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/clients/{client_id}:
    put:
      summary: Update client
      tags: [admin]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdate'
      responses:
        '200':
          description: Client updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Disable client
      tags: [admin]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Client disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ClientApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Client API key returned from /api/clients and sent as X-API-Key.
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: Admin bearer token configured via ADMIN_TOKEN.
    DisplaySecret:
      type: apiKey
      in: header
      name: X-Display-Secret
      description: Shared display secret configured via DISPLAY_SECRET.
  parameters:
    ClientId:
      name: client_id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_request:
              summary: Invalid request
              value:
                error: bad_request
                message: Missing required field
                details:
                  field: name
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_auth:
              summary: Missing or invalid auth
              value:
                error: unauthorized
                message: Invalid or missing credentials
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            not_allowed:
              summary: Not allowed
              value:
                error: forbidden
                message: You do not have access to this resource
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            not_found:
              summary: Resource not found
              value:
                error: not_found
                message: Client not found
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          examples:
            validation_error:
              summary: Validation failed
              value:
                error: validation_error
                message: Invalid payload
                details:
                  errors:
                    - field: payload_type
                      message: must be a non-empty string
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    ValidationError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
    ResponseMeta:
      type: object
      properties:
        request_id:
          type: string
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        next_page:
          type: integer
        prev_page:
          type: integer
    ClientCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        contact:
          type: string
        payload_types:
          type: array
          items:
            type: string
    ClientCreated:
      type: object
      properties:
        id:
          type: string
        api_key:
          type: string
        created_at:
          type: string
          format: date-time
      example:
        id: cli_123
        api_key: sk_live_xxx
        created_at: '2026-03-22T12:00:00Z'
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        contact:
          type: string
        payload_types:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        disabled:
          type: boolean
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        contact:
          type: string
        payload_types:
          type: array
          items:
            type: string
        disabled:
          type: boolean
      example:
        id: cli_123
        name: WeatherBot
        description: Local weather producer
        contact: ops@example.com
        payload_types: [weather]
        created_at: '2026-03-22T12:00:00Z'
        disabled: false
    PayloadSubmit:
      type: object
      required: [client_id, payload_type, data]
      properties:
        client_id:
          type: string
        payload_type:
          type: string
          enum: [raw, raw_commands, raw_pixels, simple_text_scroll, simple_text_page, rich_text_scroll, rich_text_page, billboard, clock, weather, image, animation, template, clear]
          enum: [raw, raw_commands, raw_pixels, simple_text_scroll, simple_text_page, rich_text_scroll, rich_text_page, billboard, clock, weather, image, animation, template, clear]
        template_id:
          type: string
        format_hint:
          type: string
        priority:
          type: integer
        ttl_seconds:
          type: integer
        tags:
          type: array
          items:
            type: string
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        data:
          type: object
      example:
        client_id: cli_123
        payload_type: weather
        template_id: tpl_abc
        format_hint: compact
        priority: 5
        ttl_seconds: 120
        tags: [priority, demo]
        valid_from: '2026-03-22T12:00:00Z'
        valid_to: '2026-03-22T12:05:00Z'
        data:
          temp_f: 72
          condition: Sunny
    PayloadAccepted:
      type: object
      properties:
        payload_id:
          type: string
        routed_displays:
          type: array
          items:
            type: string
        status:
          type: string
      example:
        payload_id: pld_789
        routed_displays: [disp_main]
        status: accepted
...
    PreviewSubmit:
      type: object
      required: [client_id, payload_type, data]
      properties:
        client_id:
          type: string
        payload_type:
          type: string
          enum: [raw, raw_commands, raw_pixels, simple_text_scroll, simple_text_page, rich_text_scroll, rich_text_page, billboard, clock, weather, image, animation, template, clear]
        template_id:
          type: string
        format_hint:
          type: string
        priority:
          type: integer
        ttl_seconds:
          type: integer
        tags:
          type: array
          items:
            type: string
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        data:
          type: object
        display_ids:
          type: array
          items:
            type: string
        all_displays:
          type: boolean
          default: true
      example:
        client_id: cli_123
        payload_type: simple_text_scroll
        data:
          text: Hello preview
        all_displays: true
    PreviewAccepted:
      type: object
      properties:
        preview_id:
          type: string
        routed_displays:
          type: array
          items:
            type: string
        status:
          type: string
      example:
        preview_id: prev_001
        routed_displays: [disp_main]
        status: accepted
    ValidateRequest:
      type: object
      required: [payload_type, data]
      properties:
        payload_type:
          type: string
          enum: [raw, raw_commands, raw_pixels, simple_text_scroll, simple_text_page, rich_text_scroll, rich_text_page, billboard, clock, weather, image, animation, template, clear]
        data:
          type: object
        template:
          type: string
        context:
          type: object
    ValidateResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
      example:
        valid: true
        errors: []
        warnings: []
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        payload_type:
          type: string
        template:
          type: string
        default_style:
          type: object
        created_at:
          type: string
          format: date-time
...
    TemplateList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Template'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    PayloadTypeList:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...